+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 - Support for repeating lyrics in lrc
 - `vertical_align` and `horizontal_align` to album art config, supports kitty, sixel and iterm2
 - Support for fixed Pane size in Tabs
+- Support for displaying MPD stickers in the header and queue table
+- Support for manipulation of MPD stickers via CLI
 
 ### Changed